requirements.txt
django==3.1.4
django-polymorphic==3.0.0
